+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ```min``` validation rule (float, integer)
 - ```max``` validation rule (float, integer)
 - ```regex``` validation rule (string)
+- ```list``` type
 
 ## [v0.1.0] 2019-03-28
 ### Added